add exception to Account, Budget, Categories

+++ src/main/java/com/example/personalfinance/controller/AccountController.java
@@ -4,9 +4,11 @@ import com.example.personalfinance.bean.response.AccountResponse;
 import com.example.personalfinance.bean.response.BaseResponse;
 import com.example.personalfinance.config.auth.JWTGenerator;
 import com.example.personalfinance.entity.Account;
+import com.example.personalfinance.exception.account.AccountNotFoundException;
+import com.example.personalfinance.exception.account.AccountUpdateException;
+import com.example.personalfinance.exception.InvalidInputException;
 import com.example.personalfinance.service.AccountService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,6 +21,9 @@ public class AccountController {
     private final JWTGenerator jwtGenerator;
     private final AccountService accountService;
 
+    /**
+     * Creates a new account for the authenticated user.
+     */
     @PostMapping
     public ResponseEntity<BaseResponse> createAccount(@RequestHeader(value = "Authorization") String token,
                                                       @RequestBody Account account) {
@@ -31,8 +36,16 @@ public class AccountController {
     public ResponseEntity<BaseResponse> updateAccount(@RequestHeader(value = "Authorization") String token,
                                                       @RequestBody Account account,
                                                       @RequestParam String accountId) {
-        accountService.updateAccount(account, Integer.valueOf(accountId));
-        return ResponseEntity.ok(new BaseResponse("success", accountService.getAccountById(Integer.valueOf(accountId))));
+        try {
+            accountService.updateAccount(account, Integer.valueOf(accountId));
+            return ResponseEntity.ok(new BaseResponse("success", accountService.getAccountById(Integer.valueOf(accountId))));
+        } catch (AccountNotFoundException ex) {
+            throw ex;
+        } catch (InvalidInputException ex) {
+            throw ex;
+        } catch (AccountUpdateException ex) {
+            throw ex;
+        }
     }
 
     @GetMapping
@@ -46,15 +59,10 @@ public class AccountController {
     public ResponseEntity<BaseResponse> deleteAccount(@RequestHeader(value = "Authorization") String token,
                                                       @RequestParam String accountId) {
         String userName = jwtGenerator.getUsernameFromJWT(jwtGenerator.getTokenFromHeader(token));
-        if (accountService.hasAccount(accountId)) {
-            if (accountService.hasPermission(userName, accountId)) {
-                accountService.deleteAccount(accountId);
-                return ResponseEntity.ok(new BaseResponse("success"));
-            } else {
-                return ResponseEntity.badRequest().body(new BaseResponse("couldn't delete account"));
-            }
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new BaseResponse("couldn't find account", accountId));
+        if (accountService.hasAccount(accountId) && accountService.hasPermission(userName, accountId)) {
+            accountService.deleteAccount(accountId);
+            return ResponseEntity.ok(new BaseResponse("success"));
         }
+        throw new AccountNotFoundException("Account with ID " + accountId + " not found or unauthorized access");
     }
 }
